requirements.txt
async-timeout==5.0.1
blinker==1.9.0
cassandra-driver==3.29.2
cassandra-pylib==0.0.0
click==8.1.8
faker==37.1.0
flask==3.1.0
flask-cors==5.0.1
geomet==0.2.1.post1
importlib-metadata==8.7.0
itsdangerous==2.2.0
jinja2==3.1.6
MarkupSafe==3.0.2
numpy==2.0.2
pandas==2.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
redis==6.0.0
six==1.17.0
speedtest-cli==2.1.3
tzdata==2025.2
werkzeug==3.1.3
zipp==3.21.0